requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-allauth==0.56.1
dj-rest-auth==5.0.1
python-dotenv==1.0.0
drf-spectacular[sidecar]==0.26.4
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-model-utils==4.3.1
gunicorn==21.2.0
django-allauth-ui==0.1.4
django-widget-tweaks==1.5.0
